requirements.txt
streamlit
openai
ddgs
pillow
requests